+++ android/app/src/androidTest/java/com/yhz47ow/nutritiontracker/ExampleInstrumentedTest.java
@@ -1,0 +1,18 @@
+package com.yhz47ow.nutritiontracker;
+
+import static org.junit.Assert.assertEquals;
+
+import android.content.Context;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+@RunWith(AndroidJUnit4.class)
+public class ExampleInstrumentedTest {
+    @Test
+    public void usesAppContext() {
+        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
+        assertEquals("com.yhz47ow.nutritiontracker", appContext.getPackageName());
+    }
+}