+++ android/app/src/main/java/com/yhz47ow/nutritiontracker/MainActivity.java
@@ -1,0 +1,12 @@
+package com.yhz47ow.nutritiontracker;
+
+import android.os.Bundle;
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(RestTimerPlugin.class);
+        super.onCreate(savedInstanceState);
+    }
+}